plate_worker/requirements.txt
ultralytics
paddleocr
opencv-python
redis
aio_pika
Pillow